services/langgraph_v10/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph_v10_app"
version = "0.1.0"
description = "LangGraph v10 Studio app"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph",
  "langgraph-cli[inmem]",
  "langchain",
  "langchain-openai",
  "langsmith",
  "langchain-google-genai",
  "azure-storage-blob",
  "azure-ai-documentintelligence",
  "python-dotenv"
]

[tool.setuptools.packages.find]
where = ["src"]
